owrt-qti-sepolicy:FR101595 Selinux changes to support health monitor
New daemon health monitor is created to monitor specific features and
feature is not working, need to take recovery actions as per
configuration.

Created a new cil file for new daemon health monitor to give context
to binary and all files created by it.

Handled all the denials for the functionalities

Change-Id: I776b8eb4fd0b5ce9e5ba5e02e76972b34ad4fd1a